refactor: rename captchaId to sessionId in login and captcha components to match documentation
