+++ src/app/components/forms/login-form/login-form.component.html
@@ -20,7 +20,7 @@
     </div>
   </div>
 
-  <app-captcha-input #captchaRef [captchaIdControl]="captchaId" [captchaInputControl]="captchaValue" />
+  <app-captcha-input #captchaRef [sessionIdControl]="sessionId" [captchaInputControl]="captcha" />
 
   <div class="flex flex-column gap-2 mb-5">
     <app-remember-me-input [control]="rememberMe" />
