+++ src/app/components/shared/inputs/captcha-input/captcha-input.component.html
@@ -1,6 +1,6 @@
 
 <div>
-  <input hidden  [formControl]="captchaIdControl" readonly   />
+  <input hidden  [formControl]="sessionIdControl" readonly   />
 
   <p-inputgroup styleClass="h-2-5rem">
     <p-inputgroup-addon
