+++ src/app/model/login-form-data.model.ts
@@ -2,8 +2,6 @@ export interface LoginFormData {
   username: string;
   password: string;
   rememberMe:boolean,
-  captcha: {
-    sessionId:string,
-    value:string,
-  };
+  sessionId:string,
+  captcha: string,
 }
