+++ src/app/services/toast-service/toast.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { MessageService } from 'primeng/api';
-import { IToastService } from './toast-service.interface';
+import { IToastService } from './toast.interface';
 
 @Injectable({ providedIn: 'root' })
 export class ToastService implements IToastService {
@@ -17,4 +17,5 @@ export class ToastService implements IToastService {
   info(summary: string, detail: string) {
     this.messageService.add({ severity: 'info', summary, detail });
   }
+
 }
